\name{heleramcar3-package}
\alias{heleramcar3-package}
\alias{heleramcar3}
\docType{package}
\title{
\packageTitle{heleramcar3}
}
\description{
\packageDescription{heleramcar3}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{heleramcar3}

Maintainer: \packageMaintainer{heleramcar3}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = heleramcar3)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{heleramcar3}}
%\section{Documentation Index}{\packageIndices{heleramcar3}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
